Added Marginfi Account Initialize and Passive Send Tx
Added a marginfi account initialize function as a helper to the wizard setup

Added a passive send method for transactions, because it isn't needed a agressive for this function.

Also there is a lot of code repetition in the way we send transactions currently, this should be made better in the next Commit with a better Struct? that collapse a lot of this